remove old calc bases estimator snapshots
