--- crates/sui-cost/src/lib.rs
@@ -1,4 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod estimator;
