+++ crates/sui-cost/tests/empirical_transaction_cost.rs
@@ -7,8 +7,6 @@ use sui_config::NetworkConfig;
 use sui_config::ValidatorInfo;
 use sui_core::test_utils::make_transfer_object_transaction;
 use sui_core::test_utils::make_transfer_sui_transaction;
-use sui_cost::estimator::estimate_transaction_computation_cost;
-use sui_cost::estimator::CommonTransactionCosts;
 use sui_types::base_types::SuiAddress;
 use sui_types::coin::PAY_JOIN_FUNC_NAME;
 use sui_types::coin::PAY_MODULE_NAME;
@@ -24,12 +22,43 @@ use test_utils::messages::move_transaction_with_type_tags;
 use test_utils::transaction::get_framework_object;
 use test_utils::transaction::make_publish_package;
 use test_utils::{
-    authority::{spawn_test_authorities, test_authority_configs},
+    authority::test_authority_configs,
     messages::move_transaction,
     transaction::{
         publish_counter_package, submit_shared_object_transaction, submit_single_owner_transaction,
     },
 };
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use serde::{Deserialize, Serialize};
+use strum_macros::Display;
+use strum_macros::EnumString;
+
+#[derive(
+    Debug, Eq, PartialEq, Hash, Serialize, Deserialize, Ord, PartialOrd, Clone, Display, EnumString,
+)]
+pub enum CommonTransactionCosts {
+    Publish,
+    MergeCoin,
+    SplitCoin(usize),
+    TransferWholeCoin,
+    TransferWholeSuiCoin,
+    TransferPortionSuiCoin,
+    SharedCounterCreate,
+    SharedCounterAssertValue,
+    SharedCounterIncrement,
+}
+
+impl CommonTransactionCosts {
+    pub fn is_shared_object_tx(&self) -> bool {
+        matches!(
+            self,
+            CommonTransactionCosts::SharedCounterAssertValue
+                | CommonTransactionCosts::SharedCounterIncrement
+        )
+    }
+}
 
 const TEST_DATA_DIR: &str = "tests/data/";
 
@@ -42,17 +71,11 @@ const TEST_DATA_DIR: &str = "tests/data/";
 #[tokio::test]
 async fn test_good_snapshot() -> Result<(), anyhow::Error> {
     let mut common_costs_actual: BTreeMap<String, GasCostSummary> = BTreeMap::new();
-    let mut common_costs_estimate: BTreeMap<String, GasCostSummary> = BTreeMap::new();
 
-    run_actual_and_estimate_costs()
-        .await?
-        .iter()
-        .for_each(|(k, (actual, estimate))| {
-            common_costs_actual.insert(k.clone().to_string(), actual.clone());
-            common_costs_estimate.insert(k.clone().to_string(), estimate.clone());
-        });
+    run_actual_costs().await?.iter().for_each(|(k, actual)| {
+        common_costs_actual.insert(k.clone().to_string(), actual.clone());
+    });
     assert_json_snapshot!(common_costs_actual);
-    assert_json_snapshot!(common_costs_estimate);
 
     Ok(())
 }
@@ -238,8 +261,8 @@ async fn create_txes(
     ret
 }
 
-async fn run_actual_and_estimate_costs(
-) -> Result<BTreeMap<CommonTransactionCosts, (GasCostSummary, GasCostSummary)>, anyhow::Error> {
+async fn run_actual_costs(
+) -> Result<BTreeMap<CommonTransactionCosts, GasCostSummary>, anyhow::Error> {
     let mut ret = BTreeMap::new();
     let gas_objects = generate_test_gas_objects();
     let (sender, keypair) = deterministic_random_account_key();
@@ -247,7 +270,6 @@ async fn run_actual_and_estimate_costs(
     // Get the authority configs and spawn them. Note that it is important to not drop
     // the handles (or the authorities will stop).
     let configs = test_authority_configs();
-    let handles = spawn_test_authorities(gas_objects.clone(), &configs).await;
     // Publish the move package to all authorities and get the new package ref.
     tokio::task::yield_now().await;
     tokio::time::sleep(std::time::Duration::from_secs(3)).await;
@@ -267,28 +289,7 @@ async fn run_actual_and_estimate_costs(
                 .clone()
         };
 
-        let gas_estimate = handles[0]
-            .with_async(|node| async move {
-                let state = node.state();
-                estimate_transaction_computation_cost(
-                    tx.into_inner().into_data().intent_message.value,
-                    state.clone(),
-                    None,
-                    None,
-                    None,
-                    None,
-                )
-                .await
-                .unwrap()
-            })
-            .await;
-        ret.insert(tx_type.clone(), (gas_used.clone(), gas_estimate.clone()));
-
-        // Check that the estimates are not less than actual
-        assert!(gas_used.computation_cost <= gas_estimate.computation_cost);
-        assert!(gas_used.storage_cost <= gas_estimate.storage_cost);
-        // Rebate should be less since this is returned
-        assert!(gas_used.storage_rebate >= gas_estimate.storage_rebate);
+        ret.insert(tx_type.clone(), gas_used.clone());
     }
     Ok(ret)
 }
